requirements.txt
fastapi>=0.122.0
uvicorn>=0.38.0
python-dotenv>=1.2.1
pydantic>=2.12.4
openai>=2.9.0
loguru>=0.7.3
python-multipart>=0.0.20